Norme update
